pyproject.toml
[tool.poetry]
name = "mutation-prediction"
version = "0.1.0"
description = ""
authors = ["Tilman Hoffbauer <tilman.hoffbauer@rwth-aachen.de>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
torch = "1.9.0"
scikit-learn = "^0.24.1"
numpy = "^1.19.5"
matplotlib = "^3.3.4"
glmnet = "^2.2.1"
xgboost = "^1.3.3"
tqdm = "^4.57.0"
optuna = "2.8.0"
biopython = "^1.78"
plotly = "^4.14.3"
pytorch-lightning = "^1.2.1"
hjson = "^3.0.2"
bokeh = "<2.0.0"
psycopg2 = "^2.8.6"
pandas = ">=1.2.3,<1.3.0"
redis = "^3.5.3"
transformers = "^4.5.1"
pytorch-nlp = "^0.5.0"
numba = "^0.53.1"
pyarrow = "^4.0.0"
maturin = "^0.10.6"
sentencepiece = "^0.1.96"
pyFFTW = "^0.12.0"
arch = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.6b0"
pyflakes = "^2.2.0"
isort = "^5.7.0"
notebook = "^6.2.0"
plotly = "^4.14.3"
nbdime = "^3.1.0"
snakeviz = "^2.1.0"
camelot-py = "^0.8.2"
opencv-python = "^4.5.1"
line-profiler = "^3.2.1"
pytest-randomly = "^3.5.0"
captum = "^0.3.1"
optuna-dashboard = "^0.4.1"
kaleido = "0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
